+++ src/gpu/src/TemplateMatchingCore.h
@@ -81,7 +81,7 @@ public:
 	int current_search_position;
 	int first_search_position;
 	int last_search_position;
-	long total_number_of_cccs_calculated;
+	int total_number_of_cccs_calculated;
 
 	Histogram histogram;
